+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "arcade_accelerate_rust"
+name = "arcade_accelerate"
 crate-type = ["cdylib"]
 
 [dependencies]
